basic -> advanced복붙
