+++ src/advanced/hooks/useCouponts.ts
@@ -1,0 +1,65 @@
+// TODO: 쿠폰 관리 Hook
+// 힌트:
+// 1. 쿠폰 목록 상태 관리 (localStorage 연동 고려)
+// 2. 쿠폰 추가/삭제
+//
+// 반환할 값:
+// - coupons: 쿠폰 배열
+// - addCoupon: 새 쿠폰 추가
+// - removeCoupon: 쿠폰 삭제
+
+import { useState, useEffect, useCallback } from 'react';
+
+import { initialCoupons } from '../../basic/constants';
+import { CartItem, Coupon } from '../../types';
+import { calculateCartTotal } from '../utils/calculateItem';
+
+export function useCoupons(
+  cart: CartItem[],
+  addNotification: (message: string, type?: 'error' | 'success' | 'warning') => void
+) {
+  const [coupons, setCoupons] = useState<Coupon[]>(() => {
+    const saved = localStorage.getItem('coupons');
+    if (saved) {
+      try {
+        return JSON.parse(saved);
+      } catch {
+        return initialCoupons;
+      }
+    }
+    return initialCoupons;
+  });
+  const [selectedCoupon, setSelectedCoupon] = useState<Coupon | null>(null);
+  // 쿠폰 설정
+  useEffect(() => {
+    localStorage.setItem('coupons', JSON.stringify(coupons));
+  }, [coupons]);
+
+  const applyCoupon = useCallback(
+    (coupon: Coupon) => {
+      const currentTotal = calculateCartTotal(cart, selectedCoupon).totalAfterDiscount;
+
+      if (currentTotal < 10000 && coupon.discountType === 'percentage') {
+        addNotification('percentage 쿠폰은 10,000원 이상 구매 시 사용 가능합니다.', 'error');
+        return;
+      }
+
+      setSelectedCoupon(coupon);
+      addNotification('쿠폰이 적용되었습니다.', 'success');
+    },
+    [addNotification, calculateCartTotal]
+  );
+
+  const deleteCoupon = useCallback(
+    (couponCode: string) => {
+      setCoupons((prev) => prev.filter((c) => c.code !== couponCode));
+      if (selectedCoupon?.code === couponCode) {
+        setSelectedCoupon(null);
+      }
+      addNotification('쿠폰이 삭제되었습니다.', 'success');
+    },
+    [selectedCoupon, addNotification]
+  );
+
+  return { coupons, setCoupons, setSelectedCoupon, selectedCoupon, applyCoupon, deleteCoupon };
+}
