+++ src/advanced/models/product.ts
@@ -1,0 +1,2 @@
+
+
