+++ src/advanced/utils/filterProducts.ts
@@ -1,0 +1,17 @@
+// utils/productFilter.ts
+import { Product } from '../../types';
+import { ProductWithUI } from '../App';
+
+export function filterProductsBySearchTerm(
+  products: ProductWithUI[],
+  searchTerm: string
+): Product[] {
+  if (!searchTerm) return products;
+
+  const lowered = searchTerm.toLowerCase();
+  return products.filter(
+    (product) =>
+      product.name.toLowerCase().includes(lowered) ||
+      (product.description && product.description.toLowerCase().includes(lowered))
+  );
+}
